[JOHNNIE] - 스프링 카페 6단계 - 댓글 (#51)
* feat - 글쓰기 기능 추가
1. 새로운 글 쓰기 기능
2. 전체 글 목록 조회하기 기능
3. 전체 글 중 특정 글에 대해 게시글 상세보기 기능

* feat - 회원가입 로직 추가
refactor User entity에 회원가입과 추가 요구사항을 해결 할 수 있는 loginId와 Password 필드값 추가

* - refacor : User객체에 longinId 필드를 추가했습니다 Validation들을 추가해 필수 값들도 다 받아 올 수 도 있게 했습니다.
  - User객체에 LoginId와 passWord 객체도 넣어서 회원가입시에 넣을 수 있도록 했습니다.

Feat(Login구현 추가)

:  LoginController를 만들어서 회원가입시에 입력했던 LogiD와 passWord값을 비교해서 같다면 로그인 완료 페이지로 페이지를
넘겨 줄 수 있도록 했습니다.

ModelAttribute의 값이 누락이 되거나 Valid에서 검증이 실패되면 다시 로그인 폼으로 리턴할 수 있도록 했습니다.

UserController에서 로그인한 사용자만 본인의  정보를 수정 할 수 있도록 Seesion을 사용해서 접근가능성을 설정했습니다.

* refactor[UI] UI들을 사용자 친화적이게 바꿔보았습니다.
로그인한 User가 본인의 정보를 수정하려할때 사용할수 있는 html페이지를 따로 만든 후 그 페이지에서 정보들을 받아서 프로필을 업데이트 하고 수정된 프로필로 리다이렉트 할 수 있도록 코드를 수정했습니다.

* [FEAT] DB와 통신할 Article entity 생성
2. JPA를 활용하기 위해 ArticleRepository를 생성후 JPARepository 상속받기
3. ArticleService를 JPA를 사용하면서 DB와 커넥팅

* H2 Database를 설정하고 연동했습니다.
SpringJPA를 사용해서 database와의 spl문법을 대신 했습니다.
database에 Article db 뿐만 아니라 User db도 추가했습니다.

* [Refactor] CASCASD와 orphanRemoval옵션을 통해서 도메인 계층에 비즈니스 로직을 넣어봤습니다
domain과 dto를 더욱 분리했습니다.  Article과 User를 JPA 연관관계를 줬습니다.

* [Feat] 로그인 인터셉터 만들기 - 핸들러에 넘어가기전에 로그인 체크만 하기 위해서 preHandle만을 오버라이드 해서 로그인체크를 했습니다
session값이 없거나 올바른 session데이터가 없으면 로그인폼으로 리다이렉트 시켰습니다. 이후 requestURI를 리다이렉트에 붙여 로그인후
다시 .USER가 작업하려던 페이지로 이동 시켰습니다
- 홈필드,로그인필드,회원가입필드,글 조회 기능들은 제외시켰습니다

* Login 어노테이션을 만들어서 Controller에서 반복적인 작업을 해야하는것을 제거, 이를 위해서 loginCheckArgumentResolver 생성
LoginInterceptor 에 ㄷ대한 테스트 코드 작성

* [refator]-로그인을 하면 user들을 보여주는것이 아니라 home으로 리다이렉트 할 수 있도록 하였다.

* [FEAT] 로그인한 사용자만 게시글에 대한 세부 조작 가능하도록 변경
- Interceptor를 통해 홈페이지,로그인폼,게시글 목록만 로그인안한 사용자가 접근 할 수 있도록설정
- session을통해 로그인한 사용자와 Article에 User객체를 두어 이들이 같을 경우에만 자신의 글의 세부내용과 글을 수정 삭제 가능하도록 할 수 있도록 변경

* Force merge sync with main

* Save untracked files before rebase

* [Refactor] html 파싱 폼 수정
-edit 과정에서 Putmapping으로 전송하던데이터가 파싱되지 않아 post로 임시변경

* 🔥 스프링 시큐리티 삭제, Gradle 관련 시큐리티 삭제

* ✅  Comment에 대한 domain repo를 만들었습니다

* ✅  Profile 설정  ♻️Comment에 날짜 관련 필드 BaseEntity로 빼내기

- Profile을 설정하여 dev와 local의 db를 분리했습니다
- Comment에 날짜 관련 필드를 BaseEntity 클래스로 분리하여 확정성을 늘렸습니다
- Comment과 관련된 service로직과 controller로직들을 추가했습니다

* ✅  Profile 설정  ♻️Comment에 날짜 관련 필드 BaseEntity로 빼내기
>>
>> - Profile을 설정하여 dev와 local의 db를 분리했습니다
>> - Comment에 날짜 관련 필드를 BaseEntity 클래스로 분리하여 확정성을 늘렸습니다
>> - Comment과 관련된 service로직과 controller로직들을 추가했습니다

* ♻️ 댓글을 수정시에 로그인 유저와 댓글을 쓴 유저가 같은지 검증로직 추가
view부분에서도 댓글 수정시에 수정됨을 알리는 동시에 수정된 시간을 뷰에 보일 수 있게 수정

* ✅  logout 기능 추가

* ♻️ Article과 Comment에 대해 SoftDelete 구현

-@SQLDelete와 @Where을 이용해 softDelete를 구현했습니다 또한 JPArepo를 이용해서 삭제된 데이터도 확인 할 수 있도록 구현 했습니다

* 최초 로그인시 jsessionid url 출력기능 삭제

* ♻️ Aritcle삭제 조건 추가
- Article을 삭제하려면 댓글이 없거나 댓글이 모두 Article의 작성자 소유의 댓글이여야만 한다

* ✅  회원탈퇴 기능 추가
- 회원정보창에서 회원탈퇴 기능을 추가했습니다.
- 회원 탈퇴 전에는 비밀번호 검증을 한번 더 하ㄱ도록 했습니다.
- 탈퇴시 그 User가 갖고있던 Article은 삭제된 User로 치환했습니다.

+++ src/main/java/codesquad/codestagram/controller/ArticleController.java
@@ -32,7 +32,7 @@ public class ArticleController {
     // 모든 게시글 조회
     @GetMapping()
     public String viewQuestions(Model model) {
-        List<Article> articles = articleService.getArticlesV2();
+        List<Article> articles = articleService.getArticles();
         model.addAttribute("articles", articles);
         return "qna/list";
     }
@@ -71,11 +71,11 @@ public class ArticleController {
         }
 
 
-        articleService.addArticleV2(request);
+        articleService.addArticle(request);
         return "redirect:/qna";
     }
 
-
+    //게시글 수정 입력폼노출
     @GetMapping("/articles/{articleId}/edit")
         public String showUpdateArticleForm(@Login User user, @PathVariable("articleId")Long articleId,Model model){
 
@@ -103,8 +103,27 @@ public class ArticleController {
     //게시글 삭제
     @PostMapping("/articles/{articleId}/delete")
     public String deleteArticle(@PathVariable("articleId") Long articleId) {
-        articleService.deleteArticleById(articleId);
-        return "redirect:/qna";
+        // 게시글 조회
+        Article article = articleService.findArticleById(articleId).orElseThrow(() -> new IllegalArgumentException("Article not found"));
+
+        // 댓글이 없는 경우 바로 삭제 가능
+        if (article.getComments().isEmpty()) {
+            articleService.deleteArticleById(articleId);
+            return "redirect:/qna";
+        }
+
+        // 댓글 작성자와 게시글 작성자가 모두 같은지 확인
+        boolean allCommentsByAuthor = article.getComments().stream()
+                .allMatch(comment -> comment.getUser().equals(article.getUser()));
+
+        if (allCommentsByAuthor) {
+            // 모든 댓글 작성자가 게시글 작성자인 경우 삭제 가능
+            articleService.deleteArticleById(articleId);
+            return "redirect:/qna";
+        }
+
+        // 조건에 맞지 않으면 삭제 불가능
+        return "redirect:/articles";
     }
 
 }
